# Conflicts:
#	designer/designer-backend/build.gradle
#	designer/designer-backend/src/main/java/eu.esens.espdvcd.designer/Server.java
#	designer/designer-backend/src/main/java/eu.esens.espdvcd.designer/endpoint/Endpoint.java
#	designer/designer-backend/src/main/java/eu.esens.espdvcd.designer/service/ESPDResponseToModelService.java
#	designer/designer-frontend/angular.json
#	designer/designer-frontend/package.json
#	designer/designer-frontend/src/app/app.component.ts
#	designer/designer-frontend/src/app/app.module.ts
#	designer/designer-frontend/src/app/material/material.module.ts
#	designer/designer-frontend/src/app/procedure-eo/procedure-eo.component.css
#	designer/designer-frontend/src/app/procedure-eo/procedure-eo.component.html
#	designer/designer-frontend/src/app/procedure-eo/procedure-eo.component.ts
#	designer/designer-frontend/src/app/procedure/procedure.component.html
#	designer/designer-frontend/src/app/root/root.component.html
#	designer/designer-frontend/src/app/services/apicall.service.ts
#	designer/designer-frontend/src/app/services/data.service.ts
#	designer/designer-frontend/src/app/start/start.component.html
#	designer/designer-frontend/src/app/toolbar/toolbar.component.html
#	designer/designer-frontend/src/app/welcome/welcome.component.html
#	designer/designer-frontend/src/assets/i18n/ESPD_cs.json
#	designer/designer-frontend/src/assets/i18n/ESPD_da.json
#	designer/designer-frontend/src/assets/i18n/ESPD_el.json
#	designer/designer-frontend/src/assets/i18n/ESPD_es.json
#	designer/designer-frontend/src/assets/i18n/ESPD_hr.json
#	designer/designer-frontend/src/assets/i18n/ESPD_hu.json
#	designer/designer-frontend/src/assets/i18n/ESPD_lt.json
#	designer/designer-frontend/src/assets/i18n/ESPD_lv.json
#	designer/designer-frontend/src/assets/i18n/ESPD_nl.json
#	designer/designer-frontend/src/assets/i18n/ESPD_pt.json
#	designer/designer-frontend/src/assets/i18n/ESPD_ro.json
#	designer/designer-frontend/src/assets/i18n/ESPD_sk.json
#	designer/designer-frontend/src/assets/i18n/ESPD_sl.json
#	designer/designer-frontend/src/assets/i18n/ESPD_sv.json
#	designer/designer-frontend/src/styles.css

+++ model/src/main/java/eu/esens/espdvcd/model/requirement/response/AmountResponse.java
@@ -1,6 +1,22 @@
+/**
+ * Copyright 2016-2018 University of Piraeus Research Center
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package eu.esens.espdvcd.model.requirement.response;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * Amount response
@@ -24,7 +40,7 @@ public class AmountResponse extends Response implements Serializable {
      * BusReqID: tbr92-018<br>
      * UBL syntax path: ccv:Criterion.RequirementGroup.Requirement.Response.Amount<br>
      */
-    private float amount;
+    private BigDecimal amount;
 
     /**
      * Currency
@@ -39,16 +55,15 @@ public class AmountResponse extends Response implements Serializable {
      */
     private String currency;
 
-    
     public AmountResponse() {
     }
     
-    public AmountResponse(float amount, String currency) {
+    public AmountResponse(BigDecimal amount, String currency) {
         this.amount = amount;
         this.currency = currency;
     }
 
-    public float getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
     
@@ -56,7 +71,7 @@ public class AmountResponse extends Response implements Serializable {
         return currency;
     }
 
-    public void setAmount(float amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
